+++ webservices/common/views.py
@@ -17,20 +17,25 @@ class ApiResource(utils.Resource):
     schema = None
     page_schema = None
     index_column = None
+    unique_column = None
     filter_match_fields = []
     filter_multi_fields = []
     filter_range_fields = []
+    filter_fulltext_fields = []
     query_options = []
     join_columns = {}
     aliases = {}
+    cap = 100
 
     @use_kwargs(Ref('args'))
     @marshal_with(Ref('page_schema'))
     def get(self, *args, **kwargs):
         query = self.build_query(*args, **kwargs)
+        count = counts.count_estimate(query, models.db.session, threshold=5000)
         return utils.fetch_page(
             query, kwargs,
-            model=self.model, join_columns=self.join_columns, aliases=self.aliases,
+            count=count, model=self.model, join_columns=self.join_columns, aliases=self.aliases,
+            index_column=self.index_column, cap=self.cap,
         )
 
     def build_query(self, *args, _apply_options=True, **kwargs):
@@ -38,8 +43,15 @@ class ApiResource(utils.Resource):
         query = filters.filter_match(query, kwargs, self.filter_match_fields)
         query = filters.filter_multi(query, kwargs, self.filter_multi_fields)
         query = filters.filter_range(query, kwargs, self.filter_range_fields)
+        query = self.filter_fulltext(query, kwargs)
         if _apply_options:
             query = query.options(*self.query_options)
+        return query
+
+    def filter_fulltext(self, query, kwargs):
+        for key, column in self.filter_fulltext_fields:
+            if kwargs.get(key):
+                query = utils.search_text(query, column, kwargs[key])
         return query
 
 
@@ -47,7 +59,6 @@ class ItemizedResource(ApiResource):
 
     year_column = None
     index_column = None
-    filter_fulltext_fields = []
 
     def get(self, **kwargs):
         """Get itemized resources. If multiple values are passed for `committee_id`,
@@ -66,12 +77,7 @@ class ItemizedResource(ApiResource):
             return utils.fetch_seek_page(query, kwargs, self.index_column, count=count)
         query = self.build_query(**kwargs)
         count = counts.count_estimate(query, models.db.session, threshold=5000)
-        return utils.fetch_seek_page(query, kwargs, self.index_column, count=count)
-
-    def build_query(self, **kwargs):
-        query = super().build_query(**kwargs)
-        query = self.filter_fulltext(query, kwargs)
-        return query
+        return utils.fetch_seek_page(query, kwargs, self.index_column, count=count, cap=self.cap)
 
     def join_committee_queries(self, kwargs):
         """Build and compose per-committee subqueries using `UNION ALL`.
@@ -99,9 +105,3 @@ class ItemizedResource(ApiResource):
         page_query = utils.fetch_seek_page(query, kwargs, self.index_column, count=-1, eager=False).results
         count = counts.count_estimate(query, models.db.session, threshold=5000)
         return page_query, count
-
-    def filter_fulltext(self, query, kwargs):
-        for key, column in self.filter_fulltext_fields:
-            if kwargs.get(key):
-                query = utils.search_text(query, column, kwargs[key])
-        return query